% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma_edital_com_pontos_invertidos.R
\name{soma_edital_com_pontos_invertidos}
\alias{soma_edital_com_pontos_invertidos}
\title{Essa função manipula um edital oficial para que seja possivel
somar o total com pontos de interrogação intertido}
\usage{
soma_edital_com_pontos_invertidos(edital)
}
\arguments{
\item{edital}{edital em pdf oficial, importada pelo pacote pdftools.}
}
\value{
Resultado da soma total dos créditos.
}
\description{
Essa função manipula um edital oficial para que seja possivel
somar o total com pontos de interrogação intertido
}
