% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaEdital.R
\name{warn_parse}
\alias{warn_parse}
\title{Formula de somar Valores de Créditos em Editais Oficiais}
\usage{
warn_parse()
}
\value{
retorna a seguinte mensagem "Possível problema encontrado na formatação
dos números. Utilizando estratégia mais agressiva de detecção!
Verificar resultado manualmente."
}
\description{
Essa função cria um aviso de possível problema na formatação
dos números.
}
