% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaEdital.R
\name{soma_edital}
\alias{soma_edital}
\title{Essa função manipula um edital oficial para que seja possivel
somar o total.}
\usage{
soma_edital(edital)
}
\arguments{
\item{edital}{edital em pdf oficial, importada pelo pacote pdftools.}
}
\value{
Resultado da soma total dos créditos.
}
\description{
Essa função manipula um edital oficial para que seja possivel
somar o total.
}
